[add] 削除ボタンとモーダルを作成
